requirements.txt
git+git://github.com/django/django.git#egg=django
git+git://github.com/django-mptt/django-mptt.git#egg=django-mptt
djorm-ext-core==0.4.1
djorm-ext-expressions==0.4
djorm-ext-pgarray==0.4.1
djorm-ext-pgfulltext==0.6
django-grappelli==2.4.3
django-imagekit==2.0.2
django-redis-cache==0.9.5
django-autoslug==1.6.1
django-jinja==0.8
psycopg2==2.4.5
South==0.7.6
django-model-utils==tip